+++ bundle/config/mutator/resourcemutator/apply_target_mode_test.go
@@ -176,6 +176,20 @@ func mockBundle(mode config.Mode) *bundle.Bundle {
 						},
 					},
 				},
+				DatabaseCatalogs: map[string]*resources.DatabaseCatalog{
+					"database_catalog1": {
+						DatabaseCatalog: database.DatabaseCatalog{
+							Name: "database_catalog1",
+						},
+					},
+				},
+				SyncedDatabaseTables: map[string]*resources.SyncedDatabaseTable{
+					"synced_database_table1": {
+						SyncedDatabaseTable: database.SyncedDatabaseTable{
+							Name: "synced_database_table1",
+						},
+					},
+				},
 			},
 		},
 		SyncRoot: vfs.MustNew("/Users/lennart.kats@databricks.com"),
@@ -344,7 +358,7 @@ func TestAllNonUcResourcesAreRenamed(t *testing.T) {
 				resourceType := resources.Type().Field(i).Name
 
 				// Skip resources that are not renamed
-				if resourceType == "Apps" || resourceType == "SecretScopes" || resourceType == "DatabaseInstances" {
+				if resourceType == "Apps" || resourceType == "SecretScopes" || resourceType == "DatabaseInstances" || resourceType == "DatabaseCatalogs" || resourceType == "SyncedDatabaseTables" {
 					continue
 				}