+++ bundle/deploy/terraform/pkg.go
@@ -117,6 +117,8 @@ var GroupToTerraformName = map[string]string{
 	"secret_scopes":           "databricks_secret_scope",
 	"sql_warehouses":          "databricks_sql_endpoint",
 	"database_instances":      "databricks_database_instance",
+	"database_catalogs":       "databricks_database_database_catalog",
+	"synced_database_tables":  "databricks_database_synced_database_table",
 }
 
 var TerraformToGroupName = func() map[string]string {