+++ bundle/internal/tf/schema/resource_online_store.go
@@ -9,4 +9,5 @@ type ResourceOnlineStore struct {
 	Name             string `json:"name"`
 	ReadReplicaCount int    `json:"read_replica_count,omitempty"`
 	State            string `json:"state,omitempty"`
+	WorkspaceId      string `json:"workspace_id,omitempty"`
 }