+++ bundle/terranova/pkg.go
@@ -1,0 +1,4 @@
+package terranova
+
+// How many parallel operations (API calls) are allowed
+const defaultParallelism = 10