+++ CHANGELOG.md
@@ -1,5 +1,38 @@
 # Version changelog
 
+## Release v0.266.0
+
+### Notable Changes
+* Breaking change: DABs now return an error when paths are incorrectly defined relative to the job or
+pipeline definition location instead of the configuration file location. Previously, the CLI would show a
+warning and fallback to resolving the path relative to the resource location. Users must update their bundle
+configurations to define all relative paths relative to the configuration file where the path is specified.
+See more details here: ([#3225](https://github.com/databricks/cli/pull/3225))
+* Add support volumes in Python support ([#3383])(https://github.com/databricks/cli/pull/3383))
+
+### Bundles
+* [Breaking Change] Remove deprecated path fallback mechanism for jobs and pipelines ([#3225](https://github.com/databricks/cli/pull/3225))
+* Add support for Lakebase synced database tables in DABs ([#3467](https://github.com/databricks/cli/pull/3467))
+* Rename Delta Live Tables to Lakeflow Declarative Pipelines in the default-python template ([#3476](https://github.com/databricks/cli/pull/3476)).
+* Fixed bundle init not working on Standard tier ([#3496](https://github.com/databricks/cli/pull/3496))
+
+
+## Release v0.265.0
+
+### CLI
+* Fix "cache: token not found" for auth token command ([#3447](https://github.com/databricks/cli/pull/3447))
+
+### Dependency updates
+* Upgrade TF provider to 1.87.0 ([#3430](https://github.com/databricks/cli/pull/3430))
+* Upgrade Go SDK to 0.81.0 ([#3449](https://github.com/databricks/cli/pull/3449))
+
+### Bundles
+* Add support for Lakebase database instances in DABs ([#3283](https://github.com/databricks/cli/pull/3283))
+* Add support for Lakebase database catalogs in DABs ([#3436](https://github.com/databricks/cli/pull/3436))
+* Improve error message for SDK/API errors ([#3379](https://github.com/databricks/cli/pull/3379))
+* Separate generated classes between jobs and pipelines in Python support ([#3428](https://github.com/databricks/cli/pull/3428))
+
+
 ## Release v0.264.2
 
 ### Notable Changes